server/requirements.txt
finnhub_python==2.4.19
Flask==2.3.2
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
flask_sqlalchemy==3.1.1
matplotlib==3.7.1
numpy==1.24.3
openai==1.14.3
pandas==2.2.1
scikit_learn==1.2.2
seaborn==0.13.2
setuptools==67.6.1
tensorflow==2.15.0
tensorflow_macos==2.15.0
Werkzeug==2.3.4
yfinance==0.2.33
